// src/api/generated/model/patientPaymentPlanCreate.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * DentC Backend
 * Dental PMS REST API. Column-tenant-scoped; snake_case; Orval-ready.
 * OpenAPI spec version: 1.0.0
 */
import type { PatientPaymentPlanCreatePlanType } from './patientPaymentPlanCreatePlanType';

export interface PatientPaymentPlanCreate {
  patient_id: number;
  office_id?: number | null;
  plan_bal_amt?: number | string | null;
  tx_plan_number?: string | null;
  setup_date?: string | null;
  amt_financed?: number | string | null;
  down_payment?: number | string | null;
  apr?: number | string | null;
  fin_charge?: number | string | null;
  interval_type?: string | null;
  num_payments?: number | null;
  periodic_amt?: number | string | null;
  first_due_date?: string | null;
  rem_payments?: number | null;
  rem_total_amt?: number | string | null;
  plan_type?: PatientPaymentPlanCreatePlanType;
  notes?: string | null;
  is_active?: boolean | null;
  tx_plan_amt?: number | string | null;
  treatment_plan_id?: string | null;
  billing_code?: string | null;
  financial_disclosure?: string | null;
  payment_code?: string | null;
  payment_token_id?: string | null;
  card_holder_name?: string | null;
  card_last4?: string | null;
  card_exp_month?: number | null;
  card_exp_year?: number | null;
  post_down_payment_with_card?: boolean | null;
  total_of_payments?: number | string | null;
  created_by_id?: number | null;
}